converted edge.ctrlX and Edge.ctrlY to double instead of int

+++ src/controller/GraphController.java
@@ -1558,7 +1558,7 @@ public class GraphController {
                     //add B2 to newP2 to get the newP1, the new control point
                     Point2D.Double newP1 = B2.add(newP2);
                     
-                    edge.setCtrlPoint((int) newP1.x, (int) newP1.y);
+                    edge.setCtrlPoint(newP1.x, newP1.y);
                 }
             }
             //add this vertex to the attachedVertices list
